requirements.txt
matplotlib==3.5.1
numpy==1.22.3
pandas==1.4.1
PyYAML==6.0
scikit_learn==1.0.2
seaborn==0.11.2
SQLAlchemy==1.4.32
